pyproject.toml
[project]
name = "daily-vessel-updates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.3",
    "gspread>=6.2.1",
    "openai-agents>=0.2.3",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "twilio>=9.6.5",
]
